Refactor picard.formats and picard.coverart.providers to avoid circular dependencies
