+++ picard/formats/__init__.py
@@ -4,7 +4,7 @@
 #
 # Copyright (C) 2006-2008, 2012 Lukáš Lalinský
 # Copyright (C) 2008 Will
-# Copyright (C) 2010, 2014, 2018-2019 Philipp Wolfer
+# Copyright (C) 2010, 2014, 2018-2020 Philipp Wolfer
 # Copyright (C) 2013 Michael Wiencek
 # Copyright (C) 2013, 2017-2019 Laurent Monin
 # Copyright (C) 2016-2018 Sambhav Kothari
@@ -27,7 +27,6 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
 
-from picard import log
 from picard.formats.ac3 import AC3File
 from picard.formats.apev2 import (
     AACFile,
@@ -46,6 +45,14 @@ from picard.formats.id3 import (
 )
 from picard.formats.midi import MIDIFile
 from picard.formats.mp4 import MP4File
+from picard.formats.util import (  # noqa: F401 # pylint: disable=unused-import
+    ext_to_format,
+    guess_format,
+    open_,
+    register_format,
+    supported_extensions,
+    supported_formats,
+)
 from picard.formats.vorbis import (
     FLACFile,
     OggAudioFile,
@@ -57,76 +64,6 @@ from picard.formats.vorbis import (
     OggVorbisFile,
 )
 from picard.formats.wav import WAVFile
-from picard.plugin import ExtensionPoint
-
-
-_formats = ExtensionPoint(label='formats')
-_extensions = {}
-
-
-def register_format(file_format):
-    _formats.register(file_format.__module__, file_format)
-    for ext in file_format.EXTENSIONS:
-        _extensions[ext[1:]] = file_format
-
-
-def supported_formats():
-    """Returns list of supported formats."""
-    return [(file_format.EXTENSIONS, file_format.NAME) for file_format in _formats]
-
-
-def supported_extensions():
-    """Returns list of supported extensions."""
-    return [ext for exts, name in supported_formats() for ext in exts]
-
-
-def ext_to_format(ext):
-    return _extensions.get(ext, None)
-
-
-def guess_format(filename, options=_formats):
-    """Select the best matching file type amongst supported formats."""
-    results = []
-    # Since we are reading only 128 bytes and then immediately closing the file,
-    # use unbuffered mode.
-    with open(filename, "rb", 0) as fileobj:
-        header = fileobj.read(128)
-        # Calls the score method of a particular format's associated filetype
-        # and assigns a positive score depending on how closely the fileobj's header matches
-        # the header for a particular file format.
-        results = [(option._File.score(filename, fileobj, header), option.__name__, option)
-                   for option in options
-                   if getattr(option, "_File", None)]
-    if results:
-        results.sort()
-        if results[-1][0] > 0:
-            # return the format with the highest matching score
-            return results[-1][2](filename)
-
-    # No positive score i.e. the fileobj's header did not match any supported format
-    return None
-
-
-def open_(filename):
-    """Open the specified file and return a File instance with the appropriate format handler, or None."""
-    try:
-        # Use extension based opening as default
-        i = filename.rfind(".")
-        if i >= 0:
-            ext = filename[i+1:].lower()
-            audio_file = _extensions[ext](filename)
-        else:
-            # If there is no extension, try to guess the format based on file headers
-            audio_file = guess_format(filename)
-        if not audio_file:
-            return None
-        return audio_file
-    except KeyError:
-        # None is returned if both the methods fail
-        return None
-    except Exception as error:
-        log.error("Error occurred:\n{}".format(error))
-        return None
 
 
 register_format(AACFile)
