+++ picard/formats/vorbis.py
@@ -47,11 +47,11 @@ from picard.coverart.image import (
     TagCoverArtImage,
 )
 from picard.file import File
-from picard.formats import guess_format
 from picard.formats.id3 import (
     image_type_as_id3_num,
     types_from_id3,
 )
+from picard.formats.util import guess_format
 from picard.metadata import Metadata
 from picard.util import (
     encode_filename,
